% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{apply_operator}
\alias{apply_operator}
\title{Interprets an operator from an availability condition and applies the proper
logic to the operands in that condition.}
\usage{
apply_operator(operator, operands)
}
\description{
Interprets an operator from an availability condition and applies the proper
logic to the operands in that condition.
}
\keyword{internal}
