% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{count_periods}
\alias{count_periods}
\title{Count periods}
\usage{
count_periods(start, end, period, from_zero = FALSE)
}
\arguments{
\item{start}{A vector of dates or datetimes.  Lubridate, POSIX, and text
(%Y-%m-%d or %Y-%m-%d %H:%M:%S) formats accepted.  Length must be either 1
or length(end).}

\item{end}{A vector of dates or datetimes.  Lubridate, POSIX, and text
(%Y-%m-%d or %Y-%m-%d %H:%M:%S) formats accepted.  Length must be either 1
or length(start).}

\item{period}{A single lubridate period.}

\item{from_zero}{A single logical value specifying whether to start counting
from zero (TRUE) or +1/-1 (FALSE).}
}
\value{
A numerical vector with a value for each value of \code{start}.
}
\description{
Counts the number of complete periods between two time points.  Period length
must be specified.  Timezone of \code{start[1]} is used as timezone.
}
\details{
Counting from zero yields the number of complete periods between \code{start}
and \code{end}. Counting from +1/-1 yields which period the value of
\code{end} is in.
}
