% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{eval_req}
\alias{eval_req}
\title{Takes a list-form availability requirement as input and forwards it to the
appropriate evaluation function based on the requirement's type.}
\usage{
eval_req(req, pointer, user.id, course.id)
}
\description{
Takes a list-form availability requirement as input and forwards it to the
appropriate evaluation function based on the requirement's type.
}
\keyword{internal}
