% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{eval_req_completion}
\alias{eval_req_completion}
\title{Evaluates a list-form completion availability requirement.}
\usage{
eval_req_completion(req, pointer, user.id, course.id)
}
\description{
Evaluates a list-form completion availability requirement.
}
\keyword{internal}
