% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{eval_req_date}
\alias{eval_req_date}
\title{Evaluates a list-form date availability requirement.}
\usage{
eval_req_date(req, pointer, user.id, course.id)
}
\description{
Evaluates a list-form date availability requirement.
}
\keyword{internal}
