% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{eval_req_group}
\alias{eval_req_group}
\title{Evaluates a list-form group availability requirement.}
\usage{
eval_req_group(req, pointer, user.id, course.id)
}
\description{
Evaluates a list-form group availability requirement.
}
\keyword{internal}
