% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{eval_req_grouping}
\alias{eval_req_grouping}
\title{Evaluates a list-form grouping availability requirement.}
\usage{
eval_req_grouping(req, pointer, user.id, course.id)
}
\description{
Evaluates a list-form grouping availability requirement.
}
\keyword{internal}
