% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{eval_req_profile}
\alias{eval_req_profile}
\title{Evaluates a list-form profile availability requirement.}
\usage{
eval_req_profile(req, pointer, user.id, course.id)
}
\description{
Evaluates a list-form profile availability requirement.
}
\keyword{internal}
