% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{eval_req_role}
\alias{eval_req_role}
\title{Evaluates a list-form role availability requirement.}
\usage{
eval_req_role(req, pointer, user.id, course.id)
}
\description{
Evaluates a list-form role availability requirement.
}
\keyword{internal}
