% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{extract_avail}
\alias{extract_avail}
\title{Extract availability conditions and requirements}
\usage{
extract_avail(avail, pointer)
}
\arguments{
\item{avail}{An availability condition in list format.}

\item{pointer}{An environment that will hold the information.  The
environment must have a symbol \code{reqs} that is a list.}
}
\value{
No return value.  Changes are made to \code{pointer}.
}
\description{
Extract a list of all operators, types, and keys in an availability condition
via recursive search.
}
\keyword{internal}
