% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{fetch_avail_vals}
\alias{fetch_avail_vals}
\title{Fetch user values for availability requirements}
\usage{
fetch_avail_vals(pointer, ..., con = get_session_con())
}
\arguments{
\item{pointer}{An environment that will hold the information.  The
environment must have a symbol \code{reqs} that is a list of availability
requirements from \code{extract_avail}.}

\item{...}{Potential future args.}

\item{con}{A database connection object, using the session connection by
default.}
}
\value{
No return value.  Changes are made to \code{pointer}.
}
\description{
Given a set of availability conditions, lookup users' values that can be
later used to evaluate whether those conditions were met.
}
\details{
Each Moodle component has its own types of requirements for completion.
Evaluating availability conditions requires support for each type of
component and a variety of condition operators.  The following condition
operators are supported:  \code{&}, \code{|}, \code{!&}, \code{!|}.  The
following components are supported:  completion, grade, group, grouping,
profile, and role.  The date component still needs to be incorporated.
}
\keyword{internal}
