% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certificates.R
\name{fetch_certs}
\alias{fetch_certs}
\title{Fetch certificate information}
\usage{
fetch_certs(cert_cm_id, user_ids, ..., con = get_session_con())
}
\arguments{
\item{cert_cm_id}{A single integer corresponding to a certificate course
module id (see \code{list_certs} or \code{list_cms}).}

\item{user_ids}{A vector of integers corresponding to userids.}

\item{...}{Potential future args.}

\item{con}{A database connection object, using the session connection by
default.}
}
\description{
Returns information about whether users earned and viewed a course
certificate, in addition to what datetime those events occurred.
}
\details{
Only certificate views are neatly tracked in database tables.  Whether or not
a certificate is earned has to be determined at the time this function is run
by comparing the certificate's availability conditions to trace data.  Thus,
completion results may differ between runs of this function if a course
editor changes the availability conditions of the certificate or a student
meets a requirement they had not previously met.
}
\seealso{
\code{\link{list_certs}}
}
