% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_mods.R
\name{fetch_cm}
\alias{fetch_cm}
\alias{fetch_cm_sections}
\alias{fetch_cm_instances}
\title{Fetch course module info}
\usage{
fetch_cm_sections(cm_ids, ..., con = get_session_con())

fetch_cm_instances(cm_ids, ..., con = get_session_con())
}
\arguments{
\item{cm_ids}{An integer vector containing course module ids in the database.}

\item{...}{Potential future args.}

\item{con}{A database connection object, using the session connection by
default.}
}
\value{
\code{fetch_cm_sections} returns a tibble with section information,
  one row for each element of \code{cm_ids}.

\code{fetch_cm_instances} returns a tibble with module instance
  information, one row for each element of \code{cm_ids}.
}
\description{
These functions get information about course modules from a Moodle database.
}
\seealso{
\code{\link{list_cms}}
}
\keyword{internal}
