% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment_survey.R
\name{fetch_enroll_survey}
\alias{fetch_enroll_survey}
\title{Fetch enrollment survey responses}
\usage{
fetch_enroll_survey(esurvey_id, user_ids, shape = "wide", ...,
  con = get_session_con())
}
\arguments{
\item{esurvey_id}{A single integer corresponding to an enrollment survey id
in the database.}

\item{user_ids}{A vector of integers corresponding to userids.}

\item{shape}{A single string specifying what shape the returned data should
have.  \code{"wide"} output has one row per input userid and a column for
each question.  \code{"long"} output has one row per question response and
multiple columns of information.}

\item{...}{Potential future args.}

\item{con}{A database connection object, using the session connection by
default.}
}
\value{
A tibble with enrollment survey responses.
}
\description{
Get responses to survey questions asked during course enrollment.
}
\details{
If there are multiple submissions of the same survey for the same course by
the same user, then the most recent is returned.  Not sure why there are
duplicates.  It might occur when a user edits a survey response from their
account settings, or when they click the submit button twice before the page
finishes loading.
}
\seealso{
\code{\link{list_enroll_surveys}}
}
