% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollments.R
\name{fetch_enrolled_users}
\alias{fetch_enrolled_users}
\title{Fetch users enrolled in a course}
\usage{
fetch_enrolled_users(course_id, methods = NULL, after = NULL,
  before = NULL, ..., con = get_session_con())
}
\arguments{
\item{course_id}{A single integer corresponding to a courseid in the
database.}

\item{methods}{Vector of strings specifying which enrollment methods to
include.  A NULL value allows all enrollment methods.}

\item{after}{A single UNIX datetime in any format that can be cast to a
lubridate datetime object.  Enrollments before this time will be excluded.
A NULL value will not apply this filter.}

\item{before}{A single UNIX datetime in any format that can be cast to a
lubridate datetime object.  Enrollments on and after this time will be
excluded.  A NULL value will not apply this filter.}

\item{...}{Potential future args.}

\item{con}{A database connection object, using the session connection by
default.}
}
\value{
Tibble of users enrolled in the course and enrollment details.
}
\description{
Returns the first enrollment record for each user currently enrolled in the
specified course.
}
\details{
This will not return users who were unenrolled (usually by request) from the
course.

If a user has multiple enrollments for the same course, the earlier
enrollment is returned.  An example case of duplicates is LD Fall 2017 where
the enrollment engine was used to batch assign groups, resulting in multiple
enrollment methods
}
\seealso{
\code{\link{fetch_unenrolled_users}}
}
