% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{fetch_groups}
\alias{fetch_groups}
\title{Fetch group memberships}
\usage{
fetch_groups(course_id, user_ids, shape = "wide", ...,
  con = get_session_con())
}
\arguments{
\item{course_id}{A single integer corresponding to a courseid in the
database.}

\item{user_ids}{A vector of integers corresponding to userids enrolled in the
course specified above.}

\item{shape}{A single string specifying what shape the returned data should
have.  \code{"wide"} output has one row per input userid and a column for
each possible group.  \code{"long"} output has one row per group assignment
and multiple columns of information.}

\item{...}{Potential future args.}

\item{con}{A database connection object, using the session connection by
default.}
}
\value{
A tibble with group assignment information.  Format depends on
  \code{shape} parameter.
}
\description{
Returns a list of groups to which users belong in a course.
}
