% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{fetch_questionnaires}
\alias{fetch_questionnaires}
\title{Fetch questionnaire responses}
\usage{
fetch_questionnaires(qnr_cm_id, user_ids, include_incomplete = FALSE, ...,
  con = get_session_con())
}
\arguments{
\item{qnr_cm_id}{A single integer corresponding to a questionnaire course
module id (see \code{list_cms}).}

\item{user_ids}{A vector of integers corresponding to userids.}

\item{include_incomplete}{A single logical value indicating whether to
include data from incomplete questionnaires (TRUE) or not (FALSE). Note
that a questionnaire is considered complete when the entire questionnaire
is submitted, even if a non-required question is left blank.}

\item{...}{Potential future args.}

\item{con}{A database connection object, using the session connection by
default.}
}
\value{
A tibble with questionnaire responses for each input userid.
}
\description{
Returns responses to the specified questionnaire in long format.  Incomplete
submissions can be excluded.
}
\details{
Moodle does not store short names for questionnaire questions.  If you desire
wide format data, questions will have to be manually named and reformatted.
}
\seealso{
\code{\link{list_cms}}
}
