% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{fetch_quiz_grades}
\alias{fetch_quiz_grades}
\title{Fetch quiz grades}
\usage{
fetch_quiz_grades(quiz_cm_id, user_ids, ..., con = get_session_con())
}
\arguments{
\item{quiz_cm_id}{A single integer corresponding to a quiz course module id
(see \code{list_cms}).}

\item{user_ids}{A vector of integers corresponding to userids.}

\item{...}{Potential future args.}

\item{con}{A database connection object, using the session connection by
default.}
}
\value{
A tibble with quiz grade information for each input userid.
}
\description{
Returns the most recent final quiz grades and the maximum possible grade for
the specified quiz and users.
}
\details{
Missing grades are given a value of \code{NA}.

Note that the final grades reported here are scaled, if scaling was
specified in the quiz.  After summing the individual quiz question grades,
Moodle is capable of scaling the sum.  For example, a quiz may be scored on a
5 point scale (grade sums ranging 0-5), but scaled to a range of 0-20 for
contribution to final grade in a class.  If 100 points can be earned
according to they syllabus, then a range of 0-20 would mean the quiz is 20%
of the final course grade.
}
\seealso{
\code{\link{list_cms}}
}
