% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{fetch_user_details}
\alias{fetch_user_details}
\title{Fetch user details}
\usage{
fetch_user_details(user_ids, ..., con = get_session_con())
}
\arguments{
\item{user_ids}{A vector of integers corresponding to userids.}

\item{...}{Potential future args.}

\item{con}{A database connection object, using the session connection by
default.}
}
\value{
A tibble containing user profile information.  Each row corresponds
  to a value in user_ids.
}
\description{
Gets responses to standard questions asked during Moodle account creation.
}
\details{
The timezone column has a mix of formats from new and legacy Moodle versions.
Newer Moodle versions specify the name of a city as a string.  Older versions
used a number.  A timezone of 99 signifies the Moodle server's timezone.
Other numbers signify the difference from UTC.
}
\seealso{
\code{\link{fetch_user_info}}
}
