% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{fetch_user_info}
\alias{fetch_user_info}
\title{Fetch user info}
\usage{
fetch_user_info(user_ids, ..., con = get_session_con())
}
\arguments{
\item{user_ids}{A vector of integers corresponding to userids.}

\item{...}{Potential future args.}

\item{con}{A database connection object, using the session connection by
default.}
}
\value{
A wide-format tibble with user info, where each row corresponds to a
  value of \code{user_ids}.
}
\description{
Gets responses to custom questions asked during Moodle account creation.
}
\details{
User info fields can have different response formats.  Currently,
\code{text}, \code{menu}, and \code{multiselect} are supported.

Responses to multi-select fields are reported in two ways.  A column named
after the field has a single string for each response.  The string contains
all selected values.  Each possible option is also given its own column that
has logical values indicating whether the option was selected.

Empty string values indicate no response and are changed to the R missing
value indicator \code{NA}. Do not be confused by response values that appear
similar, such as the text \code{"NA"} or \code{"N/A"}.  Non-empty strings
indicate the user selected or entered that value. Inputting a value of
\code{"NA"} is different from not responding at all, which would be coded
\code{NA}.

At the Friday Institute, some user info fields collect data that could be
used to uncover a user's identity.  This function is hard coded to only
return approved "safe" fields from which it would be difficult to uncover a
user's identity.  If a new safe field is added on a future date, it needs to
be added to the hard coded list in this package before this function will
return it.
}
\seealso{
\code{\link{fetch_user_details}}
}
