% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{json}
\alias{json}
\alias{from_JSON}
\alias{to_JSON}
\title{Convert to/from JSON}
\usage{
from_JSON(x, keep_na = TRUE, ...)

to_JSON(x, keep_na = TRUE, rm_class = TRUE, ...)
}
\arguments{
\item{x}{An object to be converted to/from JSON.}

\item{keep_na}{A single logical value indicating whether to return \code{NA}
values when received (\code{TRUE}) or not (\code{FALSE}).}

\item{...}{Additional arguments to be passed to \code{jsonlite} functions.}

\item{rm_class}{A single logical value indicating whether to remove the
\code{"json"} class from JSON objects (\code{TRUE}) or not (\code{FALSE}).}
}
\description{
These functions wrap \code{jsonlite::fromJSON} and \code{jsonlite::toJSON}
and make it easier to work with JSON inside of a data set.
}
\details{
A common use case is where all selected answers to a multi-select checkbox
survey question are stored in a single column of a wide-formatted data set.
JSON encoding is a simply way to ensure answer text doesn't interfere with
the process of parsing the different answers (e.g. if there was a semi-colon
in the answer and your separator was a semi-colon).

\code{jsonlite} does not support the preservation of \code{NA} values that
would be needed for this use case, such as when someone skipped the
multi-select question.  \code{jsonlite} throws an error reading an \code{NA}
value and writes \code{NA} to a JSON object containing an \code{NA} value,
instead of simply returning NA.  To make the functions more friendly for data
sets, \code{NA} values are returned as \code{NA} during both reading and
writing.

When storing JSON in data frames, JSON objects will be stored as a character
vector and be stripped of their \code{"json"} class that \code{jsonlite}
attaches.  Losing the \code{"json"} class does not have any detrimental
effects, but unneccessary warnings may be thrown if data frame functions
(e.g. \code{dplyr::mutate}) have to remove the class for you.  \code{to_JSON}
strips the \code{"json"} class by default to prevent warning spam.
}
\seealso{
\code{\link[jsonlite]{fromJSON}}, \code{\link[jsonlite]{toJSON}}
}
