% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{seq_to_pos}
\alias{seq_to_pos}
\title{Extract order from sequence}
\usage{
seq_to_pos(seqs)
}
\arguments{
\item{seqs}{A vector of strings, where each string specifies an ordered
sequence of cmids belonging to a course section, as found in the
course_sections database table.}
}
\value{
A tibble mapping cmids to their ordered position in their parent
  course section.  Duplicates are removed.
}
\description{
Splits a sequence string and determines the position of each cmid in the
sequence.  Returns a mapping of unique cmids to positions.
}
\keyword{internal}
