% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spread_ms}
\alias{spread_ms}
\title{Cast multi-select responses into wide format}
\usage{
spread_ms(resp, opt, resp_sep = NULL, opt_sep = NULL, prefix = "")
}
\arguments{
\item{resp}{A list of character vectors, where each character vector is a
response to a multi-select question.  Each response can be a vector of
strings, where each value represents a selected option, or a single string
of all selected options that still needs to be split (must also use
\code{sep} param).}

\item{opt}{A character vector listing all possible response options to the
multi-select question.}

\item{resp_sep}{A single string containing a regular expression.  Each string
in \code{resp} will be split where this expression matches.  A value of
NULL indicates the responses are already split.}

\item{opt_sep}{A single string containing a regular expression.  Each string
in \code{opt} will be split where this expression matches.  A value of NULL
indicates the options are already split.}

\item{prefix}{A single string to prefix each column name in the output.}
}
\value{
A tibble with one column for each possible option and one row for
  each response.  Logical values indicate whether each option was selected in
  each response.
}
\description{
Cast multi-select responses into wide format
}
\keyword{internal}
