% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{fetch_quiz_grades}
\alias{fetch_quiz_grades}
\title{Fetch quiz grades}
\usage{
fetch_quiz_grades(quiz_cm_id, user_ids, ..., con = get_session_con())
}
\arguments{
\item{quiz_cm_id}{A single integer corresponding to a quiz course module id
(see \code{list_cms}).}

\item{user_ids}{A vector of integers corresponding to userids.}

\item{...}{Potential future args.}

\item{con}{A database connection object, using the session connection by
default.}
}
\value{
A tibble with quiz grade information for each input userid.
}
\description{
Returns quiz grades and the maximum possible grade for the specified quiz and
users. Missing grades are given a value of \code{NA}.
}
\seealso{
\code{\link{list_cms}}
}
